Add unique ids to card abilities
The problem was when a card was given a temporary CONT ability, and
after that it was given another temporary non-CONT ability, and after
that CONT ability dissapeared, index of temp non-CONT is changed, but it
could be saved in a triggered ability before that. Unique id should
solve that issue.
Also there was a problem if a card had a temporary ability. If that ability
is triggered, and the card left the stage before resolving the ability,
client won't find the ability to play.