Add unique ids to cards
And use it for check timing protocol.
During check timing protocol we send an activated ability entity, which
contains the card identity for the card, whose ability is activated.
This activated ability entity is created when ability is triggered. But
when check timing starts, the card itself might be moved before, so we
hadn't a reliable way of identifying the card in this case.