Fix moving cards from view to deck
There was a problem when we were moving a card from view back to the
deck. If we were moving second to top card back to the deck, it became
top, and the rest of revealed cards were not actualy the top cards.
Which created errors with determining their position.

Now when we move from deck(view) to deck, we place the cards after
already revelaed cards.